+++ CHANGELOG.md
@@ -16,6 +16,7 @@ Features:
 * Add `Battery` module support for FreeBSD
 * Add `--disk-show-unknown` option for Disk module
 * Detect terminal version when available
+* Support `WezTerm` terminal font detection (requires [`wezterm` executable](https://wezfurlong.org/wezterm/cli/general.html) being available)
 
 Logos:
 * Raspbian (@IamNoRobot, #373)